for deploy

+++ src/Components/ListedCard/ListedCard.jsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
 import UseAuth from "../../Hooks/UseAuth";
 import { Link } from "react-router-dom";
-import {  toast } from 'react-toastify';
+// import {  toast } from 'react-toastify';
+import Swal from "sweetalert2";
 
 
 const ListedCard = () => {
@@ -20,7 +21,19 @@ const ListedCard = () => {
     // console.log(user)
    
     const handleDelete = (id) =>{
-      fetch(`http://localhost:5000/delete/${id}`,{
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          
+            
+          fetch(`http://localhost:5000/delete/${id}`,{
         method:"DELETE",
 
       })
@@ -28,10 +41,20 @@ const ListedCard = () => {
       .then (data =>{
         if (data.deletedCount > 0) {
           setControl (!control)
-			toast.error("Delete Successfully")
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your Listed Spot has been deleted.",
+            icon: "success"
+          });
 
         }
+     
       })
+
+
+        }
+      });
+      
     }
